HATCH-353: Support Decimal/Numeric

+++ packages/hatchify-core/src/dataTypes/number/getPartialOrm.ts
@@ -3,7 +3,7 @@ import type { PartialNumberORM, PartialNumberProps } from "./types"
 export function getPartialOrm(props?: PartialNumberProps): PartialNumberORM {
   return {
     sequelize: {
-      type: "DOUBLE",
+      type: "DECIMAL",
       typeArgs: [],
       allowNull: props?.required == null ? props?.required : !props.required,
       autoIncrement: props?.autoIncrement,
